refactor: 模块化重构 FBX 二进制解析器
将原本 488 行的单体 FBX-binary-parser.ts 文件拆分为 6 个功能模块：

- binary-reader.ts: 二进制数据读取器 (187行)
- property-parser.ts: 属性解析器 (62行)
- node-parser.ts: 节点解析器 (85行)
- subnode-handler.ts: 子节点处理器 (77行)
- binary-utils.ts: 工具函数 (42行)
- index.ts: 统一导出接口

重构后的主文件仅 53 行，专注于协调各模块工作。

优势：
- 单一职责原则，每个模块专注特定功能
- 提高代码可测试性和可维护性
- BinaryReader 可复用于其他二进制解析场景
- 保持完整的功能和类型安全
